chore: refactor: use APIRouter in fastapi routes, close mongo connections
