+++ env.template
@@ -11,5 +11,6 @@ OPENAI_API_KEY=
 LLAMA_CLOUD_API_KEY=
 DOCDB_NAME=
 DOCDB_COLLECTION=
-KB_CONFIG_DOCDB=
-KB_CONFIG_COLLECTION_DOCDB=
+CONFIG_DB=
+CONFIG_KB_COL=
+CONFIG_PIPELINE_COL=
